split into multiple lit test files
