--- frontend/test/lit/test_decomposition.py
@@ -1,1426 +1,0 @@
-# Copyright 2026 Xanadu Quantum Technologies Inc.
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-#     http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Tests for graph based decomposition system."""
-
-# RUN: %PYTHON %s | FileCheck %s
-
-# pylint: disable = missing-class-docstring,line-too-long,unused-argument
-
-import pennylane as qp
-from jax import numpy as jnp
-from operator2_dummy_gates import (
-    CompilableData,
-    HybridOpArg,
-    HybridWires,
-    MultiParams,
-    MultipleFullArgs,
-    MultipleRegisters,
-    NoParams,
-    NoParamsCustomOp,
-    SingleParam,
-    SingleParamCustomOp,
-    StaticData,
-    StaticDataMultiReg,
-)
-from pennylane.core import Operator2
-from pennylane.typing import Complex, Float, Int, Wire
-
-from catalyst.decomposition.decomposition_rules import (
-    compile_decomposition_rules_wrapper,
-)
-
-
-def test_compile_decomposition_rules_wrapper_entry_point():
-    """
-    Unit tests for the compile_decomposition_rules_wrapper() entry point function.
-    """
-
-    def test_to_dynamic_argnames():
-        """
-        Test that decomposing to an op with a single dynamic_argname works.
-        """
-
-        def rule_resource_fn(reg):
-            return {
-                SingleParam(x=Float, reg=Wire[2]): 2,
-                SingleParam(x=Float[2], reg=Wire[2]): 1,
-            }
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg):
-            SingleParam(x=0.1, reg=reg)
-            SingleParam(x=0.2, reg=reg)
-            SingleParam(x=jnp.array([0.3, 0.4]), reg=reg)
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, rule)
-            result = compile_decomposition_rules_wrapper(
-                "NoParams", "NoParams{}{reg:2}{}", {}, {"reg": 2}, {}
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_NoParams{}{reg:2}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "SingleParam{x:[f64,f64]}{reg:2}{}" = 1 : i64,
-    # CHECK-SAME:   "SingleParam{x:[f64]}{reg:2}{}" = 2 : i64
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:2}{}"
-    test_to_dynamic_argnames()
-
-    def test_from_dynamic_argnames():
-        """
-        Test that decomposing from an op with a single dynamic_argname works.
-        """
-
-        def rule_resource_fn(x, reg):
-            return {NoParams(reg=Wire[1]): 1}
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(x, reg):
-            NoParams(reg=reg[0])
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(SingleParam, rule)
-            result = compile_decomposition_rules_wrapper(
-                "SingleParam",
-                "SingleParam{x:[f64,f64]}{reg:2}{}",
-                {"x": ["f64", "f64"]},
-                {"reg": 2},
-                {},
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_SingleParam{x:[f64,f64]}{reg:2}{}"
-    # CHECK-SAME:   resources = {operations = {"NoParams{}{reg:1}{}" = 1 : i64}}
-    # CHECK-SAME:   target_gate = "SingleParam{x:[f64,f64]}{reg:2}{}"
-    test_from_dynamic_argnames()
-
-    def test_to_multiple_dynamic_argnames():
-        """
-        Test that decomposing to an op with multiple dynamic_argnames works.
-        """
-
-        def rule_resource_fn(reg):
-            return {
-                MultiParams(reg=Wire[1], a=Float, b=Int[2, 2], c=Complex): 1,
-            }
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg):
-            MultiParams(reg=reg, a=0.1, b=jnp.array([[1, 2], [3, 4]]), c=3 + 4j)
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, rule)
-            result = compile_decomposition_rules_wrapper(
-                "NoParams", "NoParams{}{reg:1}{}", {}, {"reg": 1}, {}
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_NoParams{}{reg:1}{}"
-    # CHECK-SAME:   resources = {operations =
-    # CHECK-SAME:   "MultiParams{a:[f64],b:{{\[\[}}i64,i64],[i64,i64{{\]\]}},c:[complex<f64>]}{reg:1}{}" = 1 : i64
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:1}{}"
-    test_to_multiple_dynamic_argnames()
-
-    def test_from_multiple_dynamic_argnames():
-        """
-        Test that decomposing from an op with multiple dynamic_argnames works.
-        """
-
-        def rule_resource_fn(reg, a, b, c):
-            return {NoParamsCustomOp(wires=Wire[2]): 1}
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg, a, b, c):
-            NoParamsCustomOp(wires=reg)
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(MultiParams, rule)
-            result = compile_decomposition_rules_wrapper(
-                "MultiParams",
-                "MultiParams{a:[f64],b:[i32,f64],c:[[i32],[f64]]}{reg:2}{}",
-                {"b": ["i32", "f64"], "c": [["i32"], ["f64"]], "a": ["f64"]},
-                {"reg": 2},
-                {},
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_MultiParams{a:[f64],b:[i32,f64],c:{{\[\[}}i32],[f64{{\]\]}}}{reg:2}{}"
-    # CHECK-SAME:   resources = {operations = {"NoParamsCustomOp{}{wires:2}{}" = 1 : i64}
-    # CHECK-SAME:   target_gate = "MultiParams{a:[f64],b:[i32,f64],c:{{\[\[}}i32],[f64{{\]\]}}}{reg:2}{}"
-    test_from_multiple_dynamic_argnames()
-
-    def test_to_multiple_wire_argnames():
-        """
-        Test that decomposing to an op with multiple wire_argnames works.
-        """
-
-        def rule_resource_fn(reg):
-            return {
-                MultipleRegisters(reg1=Wire[1], reg2=Wire[2]): 1,
-            }
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg):
-            MultipleRegisters(reg1=0, reg2=[0, 1])
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, rule)
-            result = compile_decomposition_rules_wrapper(
-                "NoParams", "NoParams{}{reg:1}{}", {}, {"reg": 1}, {}
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_NoParams{}{reg:1}{}"
-    # CHECK-SAME:   resources = {operations =
-    # CHECK-SAME:   "MultipleRegisters{}{reg1:1,reg2:2}{}" = 1 : i64
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:1}{}"
-    test_to_multiple_wire_argnames()
-
-    def test_from_multiple_wire_argnames():
-        """
-        Test that decomposing from an op with multiple wire_argnames works.
-        """
-
-        def rule_resource_fn(reg1, reg2):
-            return {NoParamsCustomOp(wires=Wire[2]): 1, NoParamsCustomOp(wires=Wire[3]): 1}
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg1, reg2):
-            NoParamsCustomOp(wires=reg1)
-            NoParamsCustomOp(wires=reg2)
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(MultipleRegisters, rule)
-            result = compile_decomposition_rules_wrapper(
-                "MultipleRegisters",
-                "MultipleRegisters{}{reg1:2,reg2:3}{}",
-                {},
-                {"reg1": 2, "reg2": 3},
-                {},
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_MultipleRegisters{}{reg1:2,reg2:3}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "NoParamsCustomOp{}{wires:2}{}" = 1 : i64
-    # CHECK-SAME:   "NoParamsCustomOp{}{wires:3}{}" = 1 : i64
-    # CHECK-SAME:   target_gate = "MultipleRegisters{}{reg1:2,reg2:3}{}"
-    test_from_multiple_wire_argnames()
-
-    def test_to_compilable_data():
-        """
-        Test that decomposing to an op with compilable_argnames works.
-        """
-
-        def rule_resource_fn(reg):
-            return {
-                CompilableData("a", "b", "thing", Wire[1]): 1,
-                CompilableData("aa", "bb", "stuff", Wire[1]): 2,
-            }
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg):
-            CompilableData(a="a", b="b", thing="thing", wires=reg[0])
-            CompilableData(a="aa", b="bb", thing="stuff", wires=reg[1])
-            CompilableData(a="aa", b="bb", thing="stuff", wires=reg[1])
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, rule)
-            result = compile_decomposition_rules_wrapper(
-                "NoParams", "NoParams{}{reg:2}{}", {}, {"reg": 2}, {}
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_NoParams{}{reg:2}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:     "CompilableData{}{wires:1}{a:a,b:b,thing:thing}" = 1 : i64,
-    # CHECK-SAME:     "CompilableData{}{wires:1}{a:aa,b:bb,thing:stuff}" = 2 : i64
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:2}{}"
-    test_to_compilable_data()
-
-    def test_from_compilable_data():
-        """
-        Test that decomposing from an op with compilable_argnames works.
-        """
-
-        def rule_resource_fn(a, b, thing, wires):
-            return {SingleParam(x=Float, reg=Wire[1]): 1}
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(a, b, thing, wires):
-            if a == 1:
-                SingleParam(x=0.1, reg=wires)
-            else:
-                SingleParam(x=1.1, reg=wires)
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(CompilableData, rule)
-            result_a1 = compile_decomposition_rules_wrapper(
-                "CompilableData",
-                "CompilableData{}{wires:1}{a:1,b:2,thing:3}",
-                {},
-                {"wires": 1},
-                {"a": 1, "b": 2, "thing": 3},
-            )
-            print(result_a1)
-
-            result_a10 = compile_decomposition_rules_wrapper(
-                "CompilableData",
-                "CompilableData{}{wires:1}{a:10,b:2,thing:3}",
-                {},
-                {"wires": 1},
-                {"a": 10, "b": 2, "thing": 3},
-            )
-            print(result_a10)
-
-    # CHECK: func.func private @"rule_CompilableData{}{wires:1}{a:1,b:2,thing:3}"
-    # CHECK-SAME:   resources = {operations = {"SingleParam{x:[f64]}{reg:1}{}" = 1 : i64}
-    # CHECK-SAME:   target_gate = "CompilableData{}{wires:1}{a:1,b:2,thing:3}"
-    # CHECK: stablehlo.constant dense<1.000000e-01> : tensor<f64>
-    # CHECK-NOT: stablehlo.constant dense<1.100000e+00> : tensor<f64>
-    #
-    # CHECK: func.func private @"rule_CompilableData{}{wires:1}{a:10,b:2,thing:3}"
-    # CHECK-SAME:   resources = {operations = {"SingleParam{x:[f64]}{reg:1}{}" = 1 : i64}
-    # CHECK-SAME:   target_gate = "CompilableData{}{wires:1}{a:10,b:2,thing:3}"
-    # CHECK: stablehlo.constant dense<1.100000e+00> : tensor<f64>
-    # CHECK-NOT: stablehlo.constant dense<1.000000e-01> : tensor<f64>
-    test_from_compilable_data()
-
-    def test_to_static_data():
-        """
-        Test that decomposing to an op with static_argnames works.
-        """
-
-        def rule_resource_fn(reg):
-            return {
-                StaticData(label="hello", reg=Wire[1]): 1,
-                StaticData(label="world", reg=Wire[1]): 2,
-            }
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg):
-            StaticData(label="hello", reg=reg[0])
-            StaticData(label="world", reg=reg[1])
-            StaticData(label="world", reg=reg[0])
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, rule)
-            result = compile_decomposition_rules_wrapper(
-                "NoParams", "NoParams{}{reg:2}{}", {}, {"reg": 2}, {}
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_NoParams{}{reg:2}{}"
-    # CHECK-DAG: "StaticData{}{reg:1}{}[[[uid_1:[0-9]+]]]" = 1
-    # CHECK-DAG: "StaticData{}{reg:1}{}[[[uid_2:[0-9]+]]]" = 2
-    # CHECK-DAG:   target_gate = "NoParams{}{reg:2}{}"
-    # CHECK: "qref.operator"({{%.+}}) {UID = [[uid_1]]
-    # CHECK: "qref.operator"({{%.+}}) {UID = [[uid_2]]
-    # CHECK: "qref.operator"({{%.+}}) {UID = [[uid_2]]
-    test_to_static_data()
-
-    def test_from_static_data():
-        """
-        Test that decomposing from an op with static_argnames works.
-        """
-
-        def rule_resource_fn(label, reg):
-            if label == 1234:
-                return {SingleParam(x=Float, reg=Wire[1]): 1}
-            else:
-                return {SingleParam(x=Float, reg=Wire[1]): 2}
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(label, reg):
-            if label == 1234:
-                SingleParam(x=0.1, reg=reg)
-            else:
-                SingleParam(x=1.1, reg=reg)
-                SingleParam(x=2.2, reg=reg)
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(StaticData, rule)
-            result_1234 = compile_decomposition_rules_wrapper(
-                "StaticData", "StaticData{}{reg:1}{}[1234]", {}, {"reg": 1}, {}, {"label": 1234}
-            )
-            print(result_1234)
-
-            result_4321 = compile_decomposition_rules_wrapper(
-                "StaticData", "StaticData{}{reg:1}{}[4321]", {}, {"reg": 1}, {}, {"label": 4321}
-            )
-            print(result_4321)
-
-    # CHECK: func.func private @"rule_StaticData{}{reg:1}{}[1234]"
-    # CHECK-SAME:   resources = {operations = {"SingleParam{x:[f64]}{reg:1}{}" = 1 : i64}
-    # CHECK-SAME:   target_gate = "StaticData{}{reg:1}{}[1234]"
-    # CHECK: stablehlo.constant dense<1.000000e-01> : tensor<f64>
-    #
-    # CHECK: func.func private @"rule_StaticData{}{reg:1}{}[4321]"
-    # CHECK-SAME:   resources = {operations = {"SingleParam{x:[f64]}{reg:1}{}" = 2 : i64}
-    # CHECK-SAME:   target_gate = "StaticData{}{reg:1}{}[4321]"
-    # CHECK: stablehlo.constant dense<1.100000e+00> : tensor<f64>
-    # CHECK: stablehlo.constant dense<2.200000e+00> : tensor<f64>
-    test_from_static_data()
-
-    def test_to_hybrid_wires():
-        """
-        Test that decomposing to an op with a wire-like hybrid_argnames works.
-        """
-
-        def rule_resource_fn(reg):
-            return {
-                HybridWires(cwires=[Wire[1]]): 2,
-                HybridWires(cwires=[Wire[1], [Wire[1], Wire[1]]]): 1,
-            }
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg):
-            HybridWires(cwires=[qp.wires.Wires(0)])
-            HybridWires(cwires=[qp.wires.Wires(1)])
-            HybridWires(
-                cwires=[qp.wires.Wires(reg[0]), [qp.wires.Wires(reg[1]), qp.wires.Wires(reg[2])]]
-            )
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, rule)
-            result = compile_decomposition_rules_wrapper(
-                "NoParams", "NoParams{}{reg:3}{}", {}, {"reg": 3}, {}
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_NoParams{}{reg:3}{}"
-    # CHECK-DAG: "HybridWires{}{}{}[[[uid_1:[0-9]+]]]" = 1
-    # CHECK-DAG: "HybridWires{}{}{}[[[uid_2:[0-9]+]]]" = 2
-    # CHECK-DAG:   target_gate = "NoParams{}{reg:3}{}"
-    # CHECK: "qref.operator"({{%.+}}) {UID = [[uid_2]]
-    # CHECK: "qref.operator"({{%.+}}) {UID = [[uid_2]]
-    # CHECK: "qref.operator"({{%.+}}) {UID = [[uid_1]]
-    test_to_hybrid_wires()
-
-    def test_from_hybrid_wires():
-        """
-        Test that decomposing from an op with a wire-like hybrid_argnames works.
-        """
-
-        def rule_resource_fn(cwires):
-            return {NoParams(reg=Wire[1]): 3}
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(cwires):
-            NoParams(reg=cwires[0])
-            NoParams(reg=cwires[1][0])
-            NoParams(reg=cwires[1][1])
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(HybridWires, rule)
-            result = compile_decomposition_rules_wrapper(
-                "HybridWires",
-                "HybridWires{}{}{}[3742]",
-                {},
-                {},
-                {},
-                extra_data={"cwires": [1, [2, qp.wires.Wires(3)]]},
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_HybridWires{}{}{}[3742]"
-    # CHECK-SAME:   resources = {operations = {"NoParams{}{reg:1}{}" = 3 : i64}}
-    # CHECK-SAME:   target_gate = "HybridWires{}{}{}[3742]"
-    # CHECK: stablehlo.constant dense<1> : tensor<i64>
-    # CHECK: stablehlo.constant dense<2> : tensor<i64>
-    # CHECK: stablehlo.constant dense<3> : tensor<i64>
-    test_from_hybrid_wires()
-
-    def test_to_hybrid_op():
-        """
-        Test that decomposing to an op with a non-wire-like hybrid_argnames works.
-        """
-
-        def rule_resource_fn(reg):
-            return {
-                HybridOpArg(
-                    angle=Float,
-                    op=StaticDataMultiReg(label="hello", reg=Wire[1], reg2=Wire[2], theta=Float),
-                    cwires=Wire[1],
-                    n_iters=100,
-                ): 2,
-                HybridOpArg(
-                    angle=Float,
-                    op=StaticDataMultiReg(label="hello", reg=Wire[1], reg2=Wire[2], theta=Float),
-                    cwires=Wire[1],
-                    n_iters=101,
-                ): 1,
-            }
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg):
-            HybridOpArg(
-                angle=0.1,
-                op=StaticDataMultiReg("hello", reg=[42], reg2=[11, 12], theta=0.4),
-                cwires=37,
-                n_iters=100,
-            )
-            HybridOpArg(
-                angle=0.2,
-                op=StaticDataMultiReg("hello", reg=[1], reg2=[2, 3], theta=0.2),
-                cwires=4,
-                n_iters=100,
-            )
-            HybridOpArg(
-                angle=0.3,
-                op=StaticDataMultiReg("hello", reg=[42], reg2=[11, 12], theta=0.4),
-                cwires=37,
-                n_iters=101,
-            )
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, rule)
-            result = compile_decomposition_rules_wrapper(
-                "NoParams", "NoParams{}{reg:3}{}", {}, {"reg": 3}, {}
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_NoParams{}{reg:3}{}"
-    # CHECK-DAG: "HybridOpArg{angle:[f64]}{cwires:1}{}[[[uid_1:[0-9]+]]]" = 1
-    # CHECK-DAG: "HybridOpArg{angle:[f64]}{cwires:1}{}[[[uid_2:[0-9]+]]]" = 2
-    # CHECK-DAG:   target_gate = "NoParams{}{reg:3}{}"
-    # CHECK: "qref.operator"({{%.+}}) {UID = [[uid_2]]
-    # CHECK: "qref.operator"({{%.+}}) {UID = [[uid_2]]
-    # CHECK: "qref.operator"({{%.+}}) {UID = [[uid_1]]
-    test_to_hybrid_op()
-
-    def test_from_hybrid_op():
-        """
-        Test that decomposing from an op with a non-wire-like hybrid_argnames works.
-        """
-
-        def rule_resource_fn(angle, op, cwires, n_iters):
-            return {NoParams(reg=Wire[1]): 1, op: 1}
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(angle, op, cwires, n_iters):
-            qp.apply(op)
-            NoParams(reg=cwires[0])
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(HybridOpArg, rule)
-            result = compile_decomposition_rules_wrapper(
-                "HybridOpArg",
-                "HybridOpArg{angle:[f64]}{cwires:1}{}[5678]",
-                {"angle": ["f64"]},
-                {"cwires": 1},
-                {},
-                extra_data={
-                    "op": StaticDataMultiReg("hello", reg=[1], reg2=[2, 3], theta=0.2),
-                    "n_iters": 100,
-                },
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_HybridOpArg{angle:[f64]}{cwires:1}{}[5678]"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "NoParams{}{reg:1}{}" = 1 : i64,
-    # CHECK-SAME:   "StaticDataMultiReg{theta:[f64]}{reg:1,reg2:2}{}[[[uid:[0-9]+]]]" = 1 : i64
-    # CHECK-SAME:   target_gate = "HybridOpArg{angle:[f64]}{cwires:1}{}[5678]"
-    # CHECK: "qref.operator"({{%.+}}) {UID = [[uid]] : i64, op_name = "StaticDataMultiReg"
-    test_from_hybrid_op()
-
-    def test_to_hybrid_op_nested():
-        """
-        Test that decomposing to an op with a hybrid_argnames that is a nested op works.
-        """
-
-        def rule_resource_fn(reg):
-            return {
-                HybridOpArg(
-                    angle=Float,
-                    op=HybridOpArg(
-                        angle=Float,
-                        op=StaticDataMultiReg(
-                            label="hello", reg=Wire[1], reg2=Wire[2], theta=Float
-                        ),
-                        cwires=Wire[3],
-                        n_iters=200,
-                    ),
-                    cwires=Wire[1],
-                    n_iters=100,
-                ): 1,
-            }
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg):
-            HybridOpArg(
-                angle=0.1,
-                op=HybridOpArg(
-                    angle=0.2,
-                    op=StaticDataMultiReg("hello", reg=[42], reg2=[11, 12], theta=0.4),
-                    cwires=[100, 200, 300],
-                    n_iters=200,
-                ),
-                cwires=37,
-                n_iters=100,
-            )
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, rule)
-            result = compile_decomposition_rules_wrapper(
-                "NoParams", "NoParams{}{reg:3}{}", {}, {"reg": 3}, {}
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_NoParams{}{reg:3}{}"
-    # CHECK-SAME: "HybridOpArg{angle:[f64]}{cwires:1}{}[[[uid:[0-9]+]]]" = 1
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:3}{}"
-    # CHECK: "qref.operator"({{%.+}}) {UID = [[uid]]
-    test_to_hybrid_op_nested()
-
-    def test_from_hybrid_op_nested():
-        """
-        Test that decomposing from an op with a hybrid_argnames that is a nested op works.
-        """
-
-        def rule_resource_fn(angle, op, cwires, n_iters):
-            return {NoParams(reg=Wire[1]): 1, op: 1, op.op: 1}
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(angle, op, cwires, n_iters):
-            qp.apply(op)
-            qp.apply(op.op)
-            NoParams(reg=cwires[0])
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(HybridOpArg, rule)
-            result = compile_decomposition_rules_wrapper(
-                "HybridOpArg",
-                "HybridOpArg{angle:[f64]}{cwires:1}{}[7654]",
-                {"angle": ["f64"]},
-                {"cwires": 1},
-                {},
-                extra_data={
-                    "op": HybridOpArg(
-                        angle=0.2,
-                        op=StaticDataMultiReg("hello", reg=[1], reg2=[2, 3], theta=0.2),
-                        cwires=2,
-                        n_iters=200,
-                    ),
-                    "n_iters": 100,
-                },
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_HybridOpArg{angle:[f64]}{cwires:1}{}[7654]"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "HybridOpArg{angle:[f64]}{cwires:1}{}[[[uid_outer:[0-9]+]]]" = 1 : i64,
-    # CHECK-SAME:   "NoParams{}{reg:1}{}" = 1 : i64,
-    # CHECK-SAME:   "StaticDataMultiReg{theta:[f64]}{reg:1,reg2:2}{}[[[uid_inner:[0-9]+]]]" = 1 : i64
-    # CHECK-SAME:   target_gate = "HybridOpArg{angle:[f64]}{cwires:1}{}[7654]"
-    # CHECK: "qref.operator"({{%.+}}) {UID = [[uid_outer]] : i64, op_name = "HybridOpArg"
-    # CHECK: "qref.operator"({{%.+}}) {UID = [[uid_inner]] : i64, op_name = "StaticDataMultiReg"
-    test_from_hybrid_op_nested()
-
-    def test_to_multiple_full_args_op():
-        """
-        Test that decomposing to an op with multiple names on all arg types works.
-        """
-
-        def rule_resource_fn(reg):
-            return {
-                MultipleFullArgs(
-                    reg1=Wire[1],
-                    reg2=Wire[2],
-                    angles1=Float,
-                    angles2=Float[2],
-                    pytree1=[1],
-                    pytree2=[2],
-                    op1=SingleParam(x=Float, reg=Wire[1]),
-                    op2=SingleParam(x=Int, reg=Wire[1]),
-                    hwires1=[Wire[1], Wire[1]],
-                    hwires2=[Wire[1]],
-                ): 2
-            }
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg):
-            MultipleFullArgs(
-                reg1=reg[0],
-                reg2=reg[1:3],
-                angles1=0.1,
-                angles2=jnp.array([0.1, 0.2]),
-                pytree1=[1],
-                pytree2=[2],
-                op1=SingleParam(x=0.1, reg=[reg[0]]),
-                op2=SingleParam(x=1, reg=[reg[1]]),
-                hwires1=[qp.wires.Wires(reg[0]), qp.wires.Wires(reg[1])],
-                hwires2=[qp.wires.Wires(reg[2])],
-            )
-            MultipleFullArgs(
-                reg1=reg[2],
-                reg2=reg[0:2],
-                angles1=1.2,
-                angles2=jnp.array([1.1, 1.2]),
-                pytree1=[1],
-                pytree2=[2],
-                op1=SingleParam(x=1.1, reg=[reg[1]]),
-                op2=SingleParam(x=2, reg=[reg[2]]),
-                hwires1=[qp.wires.Wires(reg[1]), qp.wires.Wires(reg[2])],
-                hwires2=[qp.wires.Wires(reg[0])],
-            )
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, rule)
-            result = compile_decomposition_rules_wrapper(
-                "NoParams", "NoParams{}{reg:3}{}", {}, {"reg": 3}, {}
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_NoParams{}{reg:3}{}"
-    # CHECK-DAG: "MultipleFullArgs{angles1:[f64],angles2:[f64,f64]}{reg1:1,reg2:2}{}[[[uid:[0-9]+]]]" = 2
-    # CHECK-DAG:   target_gate = "NoParams{}{reg:3}{}"
-    # CHECK: "qref.operator"({{%.+}}) {UID = [[uid]]
-    # CHECK: "qref.operator"({{%.+}}) {UID = [[uid]]
-    test_to_multiple_full_args_op()
-
-    def test_from_multiple_full_args_op():
-        """
-        Test that decomposing from an op with multiple names on all arg types works.
-        """
-
-        def rule_resource_fn(
-            reg1, reg2, angles1, angles2, pytree1, pytree2, op1, op2, hwires1, hwires2
-        ):
-            return {NoParams(reg=Wire[1]): 1, op1: 1, op2: 1}
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg1, reg2, angles1, angles2, pytree1, pytree2, op1, op2, hwires1, hwires2):
-            qp.apply(op1)
-            qp.apply(op2)
-            NoParams(reg=hwires1[0])
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(MultipleFullArgs, rule)
-            result = compile_decomposition_rules_wrapper(
-                "MultipleFullArgs",
-                "MultipleFullArgs{angles1:[f64],angles2:[f64,f64]}{reg1:1,reg2:2}{}[-4321]",
-                {"angles1": ["f64"], "angles2": ["f64", "f64"]},
-                {"reg1": 1, "reg2": 2},
-                {},
-                extra_data={
-                    "pytree1": [1],
-                    "pytree2": [2],
-                    "op1": SingleParam(x=1.1, reg=[1]),
-                    "op2": SingleParam(x=2, reg=[2]),
-                    "hwires1": [qp.wires.Wires(3), qp.wires.Wires(2)],
-                    "hwires2": [qp.wires.Wires(1)],
-                },
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule_MultipleFullArgs{angles1:[f64],angles2:[f64,f64]}{reg1:1,reg2:2}{}[-4321]"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "NoParams{}{reg:1}{}" = 1 : i64
-    # CHECK-SAME:   "SingleParam{x:[f64]}{reg:1}{}" = 1 : i64
-    # CHECK-SAME:   "SingleParam{x:[i64]}{reg:1}{}" = 1 : i64
-    # CHECK-SAME:   target_gate = "MultipleFullArgs{angles1:[f64],angles2:[f64,f64]}{reg1:1,reg2:2}{}[-4321]"
-    # CHECK: "qref.operator"({{%.+}}) {op_name = "SingleParam"
-    # CHECK: "qref.operator"({{%.+}}) {op_name = "SingleParam"
-    # CHECK: "qref.operator"({{%.+}}) {op_name = "NoParams"
-    test_from_multiple_full_args_op()
-
-    def test_multiple_rules():
-        """
-        Test when multiple rules are registered on an op.
-        """
-
-        def rule1_resource_fn(reg):
-            return {
-                SingleParam(x=Float, reg=Wire[1]): 1,
-            }
-
-        @qp.register_resources(rule1_resource_fn)
-        def rule1(reg):
-            SingleParam(x=0.1, reg=[reg])
-
-        def rule2_resource_fn(reg):
-            return {
-                CompilableData("a", "b", "thing", Wire[3]): 1,
-            }
-
-        @qp.register_resources(rule2_resource_fn)
-        def rule2(reg):
-            CompilableData(b="b", thing="thing", a="a", wires=[reg, reg + 1, reg + 2])
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, rule1)
-            qp.add_decomps(NoParams, rule2)
-            result = compile_decomposition_rules_wrapper(
-                "NoParams", "NoParams{}{reg:1}{}", {}, {"reg": 1}, {}
-            )
-            print(result)
-
-    # CHECK: func.func private @"rule1_NoParams{}{reg:1}{}"
-    # CHECK-SAME:   resources = {operations = {"SingleParam{x:[f64]}{reg:1}{}" = 1 : i64}}
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:1}{}"
-    # CHECK: func.func private @"rule2_NoParams{}{reg:1}{}"
-    # CHECK-SAME:   resources = {operations = {"CompilableData{}{wires:3}{a:a,b:b,thing:thing}" = 1 : i64}}
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:1}{}"
-    test_multiple_rules()
-
-    def test_for_loop():
-        """Test when the rule body has a for loop."""
-
-        class LayerRX(qp.core.Operator2):
-            dynamic_argnames = ("angles",)
-
-            def __init__(self, angles, wires):
-                super().__init__(angles, wires)
-
-        class TestRX(qp.core.Operator2):
-            dynamic_argnames = ("theta",)
-            wires_argnames = ("wires",)
-            arg_specs = {"theta": Float, "wires": Wire[1]}
-
-            def __init__(self, theta, wires):
-                super().__init__(theta, wires)
-
-        @qp.register_resources(lambda angles, wires: {TestRX(Float, Wire[1]): len(wires)})
-        def test_rule(angles, wires):
-            @qp.for_loop(len(wires))
-            def l(i):
-                TestRX(angles[i], wires[i])
-
-            l()  # pylint: disable=no-value-for-parameter
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(LayerRX, test_rule)
-
-            result = compile_decomposition_rules_wrapper(
-                "LayerRX", "TestID", {"angles": ["f64", "f64", "f64"]}, {"wires": 3}, {}
-            )
-            print(result)
-
-    # CHECK: func.func private @test_rule_TestID
-    # CHECK-SAME:   resources = {operations = {"TestRX{theta:[f64]}{wires:1}{}" = 3 : i64}}
-    # CHECK-SAME:   target_gate = "TestID"
-    # CHECK-DAG: stablehlo.constant dense<0> : tensor<i64>
-    # CHECK-DAG: stablehlo.constant dense<3> : tensor<i64>
-    # CHECK-DAG: stablehlo.constant dense<1> : tensor<i64>
-    # CHECK: scf.for
-    test_for_loop()
-
-    def test_if():
-        class TestOp(Operator2):
-            dynamic_argnames = ("flag",)
-            wire_argnames = ("wires",)
-
-            def __init__(self, flag, wires):
-                super().__init__(flag, wires)
-
-        @qp.register_resources(lambda flag, wires: {NoParams(Wire[1]): 1})
-        def if_decomp(flag, wires):
-            qp.cond(flag[0], NoParams)(wires)
-
-        qp.add_decomps(TestOp, if_decomp)
-
-        print(
-            compile_decomposition_rules_wrapper(
-                "TestOp", "IfID", {"flag": ["i1"]}, {"wires": 1}, {}
-            )
-        )
-
-    # CHECK: if_decomp
-    # CHECK: scf.if
-    test_if()
-
-    def test_while():
-        class WhileOp(Operator2):
-            dynamic_argnames = ("angle",)
-            wire_argnames = ("wires",)
-
-            def __init__(self, angle, wires):
-                super().__init__(angle, wires)
-
-        @qp.register_resources(lambda angle, wires: {SingleParamCustomOp(Float[1], Wire[1]): 1})
-        def while_decomp(angle, wires):
-            @qp.while_loop(lambda angle: angle < jnp.pi)
-            def while_body(angle):
-                return angle + 1.5
-
-            angle = while_body(angle)
-
-            SingleParamCustomOp(angle, wires)
-
-        qp.add_decomps(WhileOp, while_decomp)
-
-        print(
-            compile_decomposition_rules_wrapper(
-                "WhileOp", "whileID", {"angle": ["f64"]}, {"wires": 1}, {}
-            )
-        )
-
-    # CHECK: while_decomp
-    # CHECK: scf.while
-    test_while()
-
-
-test_compile_decomposition_rules_wrapper_entry_point()
-
-
-def test_lowering_time_rules():
-    """
-    Test that decomposition rules reachable from an operator are generated when lowering that
-    operator.
-    """
-
-    def test_one_rule():
-        """
-        Simple tests for when there is only one rule, i.e. graph looks like
-           A ---> B
-        """
-
-        def rule_resource_fn(reg):
-            return {
-                SingleParam(x=Float, reg=Wire[2]): 2,
-                SingleParam(x=Float[2], reg=Wire[2]): 1,
-            }
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg):
-            SingleParam(x=0.1, reg=reg[0:2])
-            SingleParam(x=0.2, reg=reg[0:2])
-            SingleParam(x=jnp.array([0.3, 0.4]), reg=reg[0:2])
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, rule)
-
-            def test_one_gate():
-                """
-                Test when circuit has just one gate.
-                """
-
-                @qp.qjit(capture=True, target="mlir")
-                @qp.qnode(qp.device("null.qubit", wires=3))
-                def c():
-                    NoParams(reg=[0, 1])
-                    return qp.state()
-
-                print(c.mlir)
-
-            # CHECK: func.func public @c()
-            # CHECK: qref.operator "NoParams"
-            # CHECK: func.func private @"__builtin_rule_NoParams{}{reg:2}{}"
-            # CHECK-SAME:   resources = {operations = {
-            # CHECK-SAME:   "SingleParam{x:[f64,f64]}{reg:2}{}" = 1 : i64,
-            # CHECK-SAME:   "SingleParam{x:[f64]}{reg:2}{}" = 2 : i64
-            # CHECK-SAME:   target_gate = "NoParams{}{reg:2}{}"
-            test_one_gate()
-
-            def test_multiple_gates_same_id():
-                """
-                Test that when circuit has multiple gates of the same id, the rule is only injected
-                once.
-                """
-
-                @qp.qjit(capture=True, target="mlir")
-                @qp.qnode(qp.device("null.qubit", wires=3))
-                def c():
-                    NoParams(reg=[0, 1])
-                    NoParams(reg=[0, 1])
-                    return qp.state()
-
-                print(c.mlir)
-
-            # CHECK: func.func public @c()
-            # CHECK: qref.operator "NoParams"
-            # CHECK: qref.operator "NoParams"
-            # CHECK: func.func private @"__builtin_rule_NoParams{}{reg:2}{}"
-            # CHECK-SAME:   resources = {operations = {
-            # CHECK-SAME:   "SingleParam{x:[f64,f64]}{reg:2}{}" = 1 : i64,
-            # CHECK-SAME:   "SingleParam{x:[f64]}{reg:2}{}" = 2 : i64
-            # CHECK-SAME:   target_gate = "NoParams{}{reg:2}{}"
-            # CHECK-NOT: func.func private @"__builtin_rule_NoParams{}{reg:2}{}"
-            test_multiple_gates_same_id()
-
-            def test_multiple_gates_different_ids():
-                """
-                Test when circuit has multiple gates of the same Operator 2 class but different ids,
-                multiple rules are generated.
-                """
-
-                @qp.qjit(capture=True, target="mlir")
-                @qp.qnode(qp.device("null.qubit", wires=3))
-                def c():
-                    NoParams(reg=[0, 1])
-                    NoParams(reg=[0, 1, 2])
-                    return qp.state()
-
-                print(c.mlir)
-
-            # CHECK: func.func public @c()
-            # CHECK: qref.operator "NoParams"
-            # CHECK: qref.operator "NoParams"
-            # CHECK: func.func private @"__builtin_rule_NoParams{}{reg:2}{}"
-            # CHECK-SAME:   resources = {operations = {
-            # CHECK-SAME:   "SingleParam{x:[f64,f64]}{reg:2}{}" = 1 : i64,
-            # CHECK-SAME:   "SingleParam{x:[f64]}{reg:2}{}" = 2 : i64
-            # CHECK-SAME:   target_gate = "NoParams{}{reg:2}{}"
-            # CHECK: func.func private @"__builtin_rule_NoParams{}{reg:3}{}"
-            # CHECK-SAME:   resources = {operations = {
-            # CHECK-SAME:   "SingleParam{x:[f64,f64]}{reg:2}{}" = 1 : i64,
-            # CHECK-SAME:   "SingleParam{x:[f64]}{reg:2}{}" = 2 : i64
-            # CHECK-SAME:   target_gate = "NoParams{}{reg:3}{}"
-            test_multiple_gates_different_ids()
-
-    test_one_rule()
-
-    def test_multiple_rules_same_gate():
-        """
-        Tests for when there are multiple distinct rules on the same gate, i.e. graph looks like
-                 +---> B
-                 |
-            A ---+
-                 |
-                 +---> C
-        """
-
-        @qp.register_resources(lambda reg: {SingleParam(x=Float, reg=Wire[2]): 1})
-        def rule1(reg):
-            SingleParam(x=0.1, reg=reg[0:2])
-
-        @qp.register_resources(lambda reg: {SingleParam(x=Float[2], reg=Wire[1]): 1})
-        def rule2(reg):
-            SingleParam(x=jnp.array([0.3, 0.4]), reg=reg[0])
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, rule1)
-            qp.add_decomps(NoParams, rule2)
-
-            @qp.qjit(capture=True, target="mlir")
-            @qp.qnode(qp.device("null.qubit", wires=3))
-            def c():
-                NoParams(reg=[0, 1])
-                NoParams(reg=[0, 1])
-                NoParams(reg=[0, 1, 2])
-                return qp.state()
-
-            print(c.mlir)
-
-    # CHECK: func.func public @c()
-    # CHECK: qref.operator "NoParams"
-    # CHECK: qref.operator "NoParams"
-    # CHECK: qref.operator "NoParams"
-    # CHECK: func.func private @"__builtin_rule1_NoParams{}{reg:2}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "SingleParam{x:[f64]}{reg:2}{}" = 1 : i64
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:2}{}"
-    # CHECK: func.func private @"__builtin_rule2_NoParams{}{reg:2}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "SingleParam{x:[f64,f64]}{reg:1}{}" = 1 : i64
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:2}{}"
-    #
-    # CHECK-NOT: func.func private @"__builtin_rule1_NoParams{}{reg:2}{}"
-    #
-    # CHECK: func.func private @"__builtin_rule1_NoParams{}{reg:3}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "SingleParam{x:[f64]}{reg:2}{}" = 1 : i64
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:3}{}"
-    # CHECK: func.func private @"__builtin_rule2_NoParams{}{reg:3}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "SingleParam{x:[f64,f64]}{reg:1}{}" = 1 : i64
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:3}{}"
-    test_multiple_rules_same_gate()
-
-    def test_multiple_rules_chained():
-        """
-        Tests for when one rule involves another rule, i.e. graph looks like
-           A ---> B ---> C
-        """
-
-        @qp.register_resources(lambda reg: {SingleParam(x=Float, reg=Wire[1]): 1})
-        def rule1(reg):
-            SingleParam(x=0.1, reg=reg[0])
-
-        @qp.register_resources(
-            lambda x, reg: {CompilableData(a="a", b="b", thing="thing", wires=Wire[1]): 1}
-        )
-        def rule2(x, reg):
-            CompilableData(a="a", b="b", thing="thing", wires=reg[0])
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, rule1)
-            qp.add_decomps(SingleParam, rule2)
-
-            @qp.qjit(capture=True, target="mlir")
-            @qp.qnode(qp.device("null.qubit", wires=3))
-            def c():
-                NoParams(reg=[0, 1])
-                return qp.state()
-
-            print(c.mlir)
-
-    # CHECK: func.func public @c()
-    # CHECK: qref.operator "NoParams"
-    # CHECK: func.func private @"__builtin_rule1_NoParams{}{reg:2}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "SingleParam{x:[f64]}{reg:1}{}" = 1 : i64
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:2}{}"
-    # CHECK: func.func private @"__builtin_rule2_SingleParam{x:[f64]}{reg:1}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "CompilableData{}{wires:1}{a:a,b:b,thing:thing}" = 1 : i64
-    # CHECK-SAME:   target_gate = "SingleParam{x:[f64]}{reg:1}{}"
-    test_multiple_rules_chained()
-
-    def test_multiple_rules_chained_and_branch():
-        """
-        Tests for when one rule involves another rule and then branches out, i.e. graph looks like
-                       +---> C
-                       |
-            A --- B ---+
-                       |
-                       +---> D
-        """
-
-        @qp.register_resources(lambda reg: {SingleParam(x=Float, reg=Wire[1]): 1})
-        def ruleAB(reg):
-            SingleParam(x=0.1, reg=reg[0])
-
-        @qp.register_resources(
-            lambda x, reg: {CompilableData(a="a", b="b", thing="thing", wires=Wire[1]): 1}
-        )
-        def ruleBC(x, reg):
-            CompilableData(a="a", b="b", thing="thing", wires=reg[0])
-
-        @qp.register_resources(
-            lambda x, reg: {CompilableData(a="alpha", b="beta", thing="stuff", wires=Wire[1]): 1}
-        )
-        def ruleBD(x, reg):
-            CompilableData(a="alpha", b="beta", thing="stuff", wires=reg[0])
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, ruleAB)
-            qp.add_decomps(SingleParam, ruleBC)
-            qp.add_decomps(SingleParam, ruleBD)
-
-            @qp.qjit(capture=True, target="mlir")
-            @qp.qnode(qp.device("null.qubit", wires=3))
-            def c():
-                NoParams(reg=[0, 1])
-                return qp.state()
-
-            print(c.mlir)
-
-    # CHECK: func.func public @c()
-    # CHECK: qref.operator "NoParams"
-    # CHECK: func.func private @"__builtin_ruleAB_NoParams{}{reg:2}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "SingleParam{x:[f64]}{reg:1}{}" = 1 : i64
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:2}{}"
-    # CHECK: func.func private @"__builtin_ruleBC_SingleParam{x:[f64]}{reg:1}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "CompilableData{}{wires:1}{a:a,b:b,thing:thing}" = 1 : i64
-    # CHECK-SAME:   target_gate = "SingleParam{x:[f64]}{reg:1}{}"
-    # CHECK: func.func private @"__builtin_ruleBD_SingleParam{x:[f64]}{reg:1}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "CompilableData{}{wires:1}{a:alpha,b:beta,thing:stuff}" = 1 : i64
-    # CHECK-SAME:   target_gate = "SingleParam{x:[f64]}{reg:1}{}"
-    test_multiple_rules_chained_and_branch()
-
-    def test_with_cycles():
-        """
-        Tests for when the decomposition rules have cycles, i.e. graph looks like
-              A
-              |
-           +--+--+
-           |     ^
-           v     |
-           B---> C
-        """
-
-        @qp.register_resources(lambda reg: {SingleParam(x=Float, reg=Wire[1]): 1})
-        def ruleAB(reg):
-            SingleParam(x=0.1, reg=reg[0])
-
-        @qp.register_resources(
-            lambda x, reg: {CompilableData(a="a", b="b", thing="thing", wires=Wire[1]): 1}
-        )
-        def ruleBC(x, reg):
-            CompilableData(a="a", b="b", thing="thing", wires=reg[0])
-
-        @qp.register_resources(lambda a, b, thing, wires: {NoParams(reg=Wire[2]): 1})
-        def ruleCA(a, b, thing, wires):
-            NoParams(reg=[0, 1])
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, ruleAB)
-            qp.add_decomps(SingleParam, ruleBC)
-            qp.add_decomps(CompilableData, ruleCA)
-
-            @qp.qjit(capture=True, target="mlir")
-            @qp.qnode(qp.device("null.qubit", wires=3))
-            def c():
-                NoParams(reg=[0, 1])
-                NoParams(reg=[0, 1, 2])
-                return qp.state()
-
-            print(c.mlir)
-
-    # CHECK: func.func public @c()
-    # CHECK: qref.operator "NoParams"
-    # CHECK: func.func private @"__builtin_ruleAB_NoParams{}{reg:2}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "SingleParam{x:[f64]}{reg:1}{}" = 1 : i64
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:2}{}"
-    # CHECK: func.func private @"__builtin_ruleBC_SingleParam{x:[f64]}{reg:1}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "CompilableData{}{wires:1}{a:a,b:b,thing:thing}" = 1 : i64
-    # CHECK-SAME:   target_gate = "SingleParam{x:[f64]}{reg:1}{}"
-    # CHECK: func.func private @"__builtin_ruleCA_CompilableData{}{wires:1}{a:a,b:b,thing:thing}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "NoParams{}{reg:2}{}" = 1 : i64
-    # CHECK-SAME:   target_gate = "CompilableData{}{wires:1}{a:a,b:b,thing:thing}"
-    # CHECK: func.func private @"__builtin_ruleAB_NoParams{}{reg:3}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "SingleParam{x:[f64]}{reg:1}{}" = 1 : i64
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:3}{}"
-    test_with_cycles()
-
-    def test_to_multiple_full_args_op():
-        """
-        Test that decomposing to an op with multiple names on all arg types works.
-        """
-
-        def rule_resource_fn(reg):
-            return {
-                MultipleFullArgs(
-                    reg1=Wire[1],
-                    reg2=Wire[2],
-                    angles1=Float,
-                    angles2=Float[2],
-                    pytree1=[1],
-                    pytree2=[2],
-                    op1=SingleParam(x=Float, reg=Wire[1]),
-                    op2=SingleParam(x=Int, reg=Wire[1]),
-                    hwires1=[Wire[1], Wire[1]],
-                    hwires2=[Wire[1]],
-                ): 2
-            }
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg):
-            MultipleFullArgs(
-                reg1=reg[0],
-                reg2=reg[1:3],
-                angles1=0.1,
-                angles2=jnp.array([0.1, 0.2]),
-                pytree1=[1],
-                pytree2=[2],
-                op1=SingleParam(x=0.1, reg=[reg[0]]),
-                op2=SingleParam(x=1, reg=[reg[1]]),
-                hwires1=[qp.wires.Wires(reg[0]), qp.wires.Wires(reg[1])],
-                hwires2=[qp.wires.Wires(reg[2])],
-            )
-            MultipleFullArgs(
-                reg1=reg[2],
-                reg2=reg[0:2],
-                angles1=1.2,
-                angles2=jnp.array([1.1, 1.2]),
-                pytree1=[1],
-                pytree2=[2],
-                op1=SingleParam(x=1.1, reg=[reg[1]]),
-                op2=SingleParam(x=2, reg=[reg[2]]),
-                hwires1=[qp.wires.Wires(reg[1]), qp.wires.Wires(reg[2])],
-                hwires2=[qp.wires.Wires(reg[0])],
-            )
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, rule)
-
-            @qp.qjit(capture=True, target="mlir")
-            @qp.qnode(qp.device("null.qubit", wires=3))
-            def c():
-                NoParams(reg=[0, 1, 2])
-                return qp.state()
-
-            print(c.mlir)
-
-    # CHECK: func.func public @c()
-    # CHECK: qref.operator "NoParams"
-    # CHECK: func.func private @"__builtin_rule_NoParams{}{reg:3}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "MultipleFullArgs{angles1:[f64],angles2:[f64,f64]}{reg1:1,reg2:2}{}[[[uid:[0-9]+]]]" = 2 : i64
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:3}{}"
-    # CHECK: qref.operator "MultipleFullArgs"
-    # CHECK-NEXT:  UID([[uid]]
-    # CHECK: qref.operator "MultipleFullArgs"
-    # CHECK-NEXT:  UID([[uid]]
-    test_to_multiple_full_args_op()
-
-    def test_from_multiple_full_args_op():
-        """
-        Test that decomposing from an op with multiple names on all arg types works.
-        """
-
-        def rule_resource_fn(
-            reg1, reg2, angles1, angles2, pytree1, pytree2, op1, op2, hwires1, hwires2
-        ):
-            return {NoParams(reg=Wire[1]): 2}
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg1, reg2, angles1, angles2, pytree1, pytree2, op1, op2, hwires1, hwires2):
-            NoParams(reg=0)
-            NoParams(reg=0)
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(MultipleFullArgs, rule)
-
-            @qp.qjit(capture=True, target="mlir")
-            @qp.qnode(qp.device("null.qubit", wires=3))
-            def c():
-                MultipleFullArgs(
-                    reg1=1,
-                    reg2=2,
-                    angles1=0.1,
-                    angles2=0.2,
-                    pytree1=[],
-                    pytree2=[],
-                    op1=SingleParam(x=1.1, reg=[0]),
-                    op2=SingleParam(x=2, reg=[1]),
-                    hwires1=[qp.wires.Wires(1), qp.wires.Wires(2)],
-                    hwires2=[qp.wires.Wires(0)],
-                )
-                return qp.state()
-
-            print(c.mlir)
-
-    # CHECK: func.func public @c()
-    # CHECK: qref.operator "MultipleFullArgs"
-    # CHECK: UID([[uid:[0-9]+]])
-    # CHECK: func.func private @"__builtin_rule_MultipleFullArgs{angles1:[f64],angles2:[f64]}{reg1:1,reg2:1}{}[[[uid]]]"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "NoParams{}{reg:1}{}" = 2 : i64
-    # CHECK-SAME:   target_gate = "MultipleFullArgs{angles1:[f64],angles2:[f64]}{reg1:1,reg2:1}{}[[[uid]]]"
-    test_from_multiple_full_args_op()
-
-    def test_to_custom_op():
-        """
-        Test that decomposing to a custom op works.
-        """
-
-        def rule_resource_fn(reg):
-            return {SingleParamCustomOp(x=Float, wires=Wire[1]): 1}
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(reg):
-            SingleParamCustomOp(x=0.1, wires=reg[0])
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(NoParams, rule)
-
-            @qp.qjit(capture=True, target="mlir")
-            @qp.qnode(qp.device("null.qubit", wires=3))
-            def c():
-                NoParams(reg=[0, 1, 2])
-                return qp.state()
-
-            print(c.mlir)
-
-    # CHECK: func.func public @c()
-    # CHECK: qref.operator "NoParams"
-    # CHECK: func.func private @"__builtin_rule_NoParams{}{reg:3}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "SingleParamCustomOp{x:[f64]}{wires:1}{}" = 1 : i64
-    # CHECK-SAME:   target_gate = "NoParams{}{reg:3}{}"
-    # CHECK: qref.custom "SingleParamCustomOp"
-    test_to_custom_op()
-
-    def test_from_custom_op():
-        """
-        Test that decomposing from a custom op works.
-        """
-
-        def rule_resource_fn(x, wires):
-            return {NoParams(reg=Wire[1]): 1}
-
-        @qp.register_resources(rule_resource_fn)
-        def rule(x, wires):
-            NoParams(reg=0)
-
-        with qp.decomposition.local_decomps():
-            qp.add_decomps(SingleParamCustomOp, rule)
-
-            @qp.qjit(capture=True, target="mlir")
-            @qp.qnode(qp.device("null.qubit", wires=3))
-            def c():
-                SingleParamCustomOp(x=0.1, wires=[0, 1])
-                return qp.state()
-
-            print(c.mlir)
-
-    # CHECK: func.func public @c()
-    # CHECK: qref.custom "SingleParamCustomOp"
-    # CHECK: func.func private @"__builtin_rule_SingleParamCustomOp{x:[f64]}{wires:2}{}"
-    # CHECK-SAME:   resources = {operations = {
-    # CHECK-SAME:   "NoParams{}{reg:1}{}" = 1 : i64
-    # CHECK-SAME:   target_gate = "SingleParamCustomOp{x:[f64]}{wires:2}{}"
-    # CHECK: qref.operator "NoParams"
-    test_from_custom_op()
-
-    def test_multirz():
-        """
-        Test that decomposing qp.MultiRZ works.
-        """
-
-        @qp.qjit(capture=True, target="mlir")
-        @qp.qnode(qp.device("null.qubit", wires=3))
-        def c():
-            qp.MultiRZ(theta=0.1, wires=[0, 1])
-            qp.MultiRZ(theta=0.2, wires=[2])
-            return qp.state()
-
-        print(c.mlir)
-
-    # CHECK: func.func public @c()
-    # CHECK: qref.multirz({{%.+}}) {{%.+}}, {{%.+}} : !qref.bit, !qref.bit
-    # CHECK: qref.multirz({{%.+}}) {{%.+}} : !qref.bit
-    # CHECK: func.func private @"__builtin__multi_rz_decomposition_MultiRZ{theta:[f64]}{wires:2}{}"
-    # CHECK-SAME:   target_gate = "MultiRZ{theta:[f64]}{wires:2}{}"
-    # CHECK: func.func private @"__builtin__multi_rz_decomposition_MultiRZ{theta:[f64]}{wires:1}{}"
-    # CHECK-SAME:   target_gate = "MultiRZ{theta:[f64]}{wires:1}{}"
-    test_multirz()
-
-    def test_paulirot():
-        """
-        Test that decomposing qp.PauliRot works.
-        """
-
-        @qp.qjit(capture=True, target="mlir")
-        @qp.qnode(qp.device("null.qubit", wires=3))
-        def c():
-            qp.PauliRot(theta=0.1, pauli_word="XX", wires=[0, 1])
-            qp.PauliRot(theta=0.2, pauli_word="Z", wires=[2])
-            return qp.state()
-
-        print(c.mlir)
-
-    # CHECK: func.func public @c()
-    # CHECK: qref.paulirot ["X", "X"]({{%.+}}) {{%.+}}, {{%.+}} : !qref.bit, !qref.bit
-    # CHECK: qref.paulirot ["Z"]({{%.+}}) {{%.+}} : !qref.bit
-    # CHECK: func.func private @"__builtin__pauli_rot_decomposition_PauliRot{theta:[f64]}{wires:2}{pauli_word:XX}"
-    # CHECK-SAME:   target_gate = "PauliRot{theta:[f64]}{wires:2}{pauli_word:XX}"
-    # CHECK: func.func private @"__builtin__pauli_rot_decomposition_PauliRot{theta:[f64]}{wires:1}{pauli_word:Z}"
-    # CHECK-SAME:   target_gate = "PauliRot{theta:[f64]}{wires:1}{pauli_word:Z}"
-    test_paulirot()
-
-
-test_lowering_time_rules()
