+++ frontend/test/lit/test_operator2/test_lowering.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Tests for operator in Catalyst."""
+"""Tests for lowering PennyLane Operator2 classes in Catalyst."""
 
 # pylint: disable = missing-function-docstring,line-too-long
